move sensors to pypms
