+++ tests/sensor/test_extra_commands.py
@@ -3,8 +3,8 @@ import os
 import pytest
 
 os.environ["LEVEL"] = "DEBUG"
-import pms.sensor.honeywell.extra_commands as HPMA
-import pms.sensor.novafitness.extra_commands as SDS
+import pypms.honeywell.extra_commands as HPMA
+import pypms.novafitness.extra_commands as SDS
 
 
 @pytest.mark.parametrize(
