+++ tests/sensor/test_obsdata.py
@@ -5,11 +5,11 @@ from dataclasses import asdict
 import pytest
 
 os.environ["LEVEL"] = "DEBUG"
-from pms.sensor.bosch_sensortec import mcu680
-from pms.sensor.honeywell import hpma115c0, hpma115s0
-from pms.sensor.novafitness import sds01x, sds198
-from pms.sensor.plantower import pms5003st, pms5003t, pmsx003
-from pms.sensor.senserion import sps30
+from pypms.bosch_sensortec import mcu680
+from pypms.honeywell import hpma115c0, hpma115s0
+from pypms.novafitness import sds01x, sds198
+from pypms.plantower import pms5003st, pms5003t, pmsx003
+from pypms.senserion import sps30
 
 
 @pytest.mark.parametrize("fmt", "header csv pm num cf raw error".split())
